+++ src/main/java/gov/usgs/cida/qw/springinit/OAuth2ResourceServerConfig.java
@@ -1,16 +1,30 @@
 package gov.usgs.cida.qw.springinit;
 
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
 import org.springframework.security.oauth2.config.annotation.web.configuration.EnableResourceServer;
 import org.springframework.security.oauth2.config.annotation.web.configuration.ResourceServerConfigurerAdapter;
+import org.springframework.security.oauth2.config.annotation.web.configurers.ResourceServerSecurityConfigurer;
+import org.springframework.security.oauth2.provider.token.DefaultTokenServices;
+import org.springframework.security.oauth2.provider.token.TokenStore;
 
 @Configuration
 @EnableResourceServer
 @Profile("internal")
 public class OAuth2ResourceServerConfig extends ResourceServerConfigurerAdapter {
+
+	@Value("${security.oauth2.resource.id}")
+	private String resourceId;
+
+	@Autowired
+	private DefaultTokenServices tokenServices;
+
+	@Autowired
+	private TokenStore tokenStore;
 
 	@Override
 	public void configure(HttpSecurity http) throws Exception {
@@ -28,4 +42,11 @@ public class OAuth2ResourceServerConfig extends ResourceServerConfigurerAdapter 
 		;
 	}
 
+	@Override
+	public void configure(ResourceServerSecurityConfigurer resources) {
+		//This is required to enable limiting by resourceId.
+		resources.resourceId(resourceId)
+			.tokenServices(tokenServices)
+			.tokenStore(tokenStore);
+	}
 }